refactor: split settings.rs into settings/ directory with models.rs (#127 step 7)
Convert settings.rs (1930 lines) into settings/ directory:
- settings/models.rs: all struct/enum definitions (~789 lines)
- settings/mod.rs: IO, migration, memo functions + pub use models::*

All existing API preserved via re-exports. 393 tests passing.
